Refactor velocity task structure
Move velocity task from nested locomotion directory to flatter structure:
  - src/mjlab/tasks/locomotion/velocity/ → src/mjlab/tasks/velocity/
  - Update all import paths in scripts and configurations
