+++ scripts/velocity/rl/play.py
@@ -9,11 +9,11 @@ from typing_extensions import assert_never
 
 from mjlab.rl import RslRlVecEnvWrapper
 from mjlab.rl.config import RslRlOnPolicyRunnerCfg
-from mjlab.tasks.locomotion.velocity.rl import (
+from mjlab.tasks.velocity.rl import (
   attach_onnx_metadata,
   export_velocity_policy_as_onnx,
 )
-from mjlab.tasks.locomotion.velocity.velocity_env_cfg import (
+from mjlab.tasks.velocity.velocity_env_cfg import (
   LocomotionVelocityEnvCfg,
 )
 from mjlab.third_party.isaaclab.isaaclab_tasks.utils.parse_cfg import (
