--- src/mjlab/tasks/locomotion/__init__.py
@@ -1,1 +1,0 @@
-"""Locomotion environments for legged robots."""
