+++ src/mjlab/tasks/velocity/__init__.py
@@ -1,0 +1,1 @@
+"""Velocity tracking environments for legged robots."""
