+++ src/mjlab/tasks/velocity/config/g1/rough_env_cfg.py
@@ -4,7 +4,7 @@ from mjlab.asset_zoo.robots.unitree_g1.g1_constants import (
   G1_ACTION_SCALE,
   G1_ROBOT_CFG,
 )
-from mjlab.tasks.locomotion.velocity.velocity_env_cfg import (
+from mjlab.tasks.velocity.velocity_env_cfg import (
   LocomotionVelocityEnvCfg,
 )
