+++ src/mjlab/tasks/velocity/rl/__init__.py
@@ -1,8 +1,8 @@
-from mjlab.tasks.locomotion.velocity.rl.exporter import (
+from mjlab.tasks.velocity.rl.exporter import (
   attach_onnx_metadata,
   export_velocity_policy_as_onnx,
 )
-from mjlab.tasks.locomotion.velocity.rl.runner import VelocityOnPolicyRunner
+from mjlab.tasks.velocity.rl.runner import VelocityOnPolicyRunner
 
 __all__ = [
   "VelocityOnPolicyRunner",
